Create viewModel for AddURLView
