+++ brain-marks/Add/AddURLView.swift
@@ -15,7 +15,9 @@ struct AddURLView: View {
     @Environment(\.presentationMode) var presentationMode
     let categories: [AWSCategory]
     
-    let pasteBoard    = UIPasteboard.general
+    @StateObject var viewModel = AddURLViewModel()
+    
+    let pasteBoard = UIPasteboard.general
     
     var body: some View {
         NavigationView {
@@ -46,7 +48,7 @@ struct AddURLView: View {
                         print("Please select a CATEGORY")
                         showingAlert = true
                     } else {
-                        get(url: newEntry) { result in
+                        viewModel.fetchTweet(url: newEntry) { result in
                             switch result {
                             case .success(let tweet):
                                 
@@ -77,57 +79,5 @@ struct AddURLView: View {
                   message: Text("You must select a category"),
                   dismissButton: .default(Text("Ok")))
         }
-    }
-}
-
-extension AddURLView {
-    func get(url:String, completion: @escaping (Result<ReturnedTweet, Error>) -> Void) {
-        
-        let id = url.components(separatedBy: "/").last!.components(separatedBy: "?")[0]
-        var request = URLRequest(url: URL(string: "https://api.twitter.com/2/tweets/\(id)?expansions=author_id")!,
-                                 timeoutInterval: Double.infinity)
-        
-        request.addValue("Bearer \(Secrets.bearerToken)", forHTTPHeaderField: "Authorization")
-        
-        request.httpMethod = "GET"
-        
-        let task = URLSession.shared.dataTask(with: request) { data, response, error in
-            guard let data = data else {
-                print(String(describing: error))
-                
-                return
-            }
-            
-            if let response = response as? HTTPURLResponse {
-                guard (200 ... 299) ~= response.statusCode else {
-                    print("Status code :- \(response.statusCode)")
-                    return
-                }
-                
-                do {
-                    if error == nil {
-                        let result = try JSONDecoder().decode(Response.self, from: data)
-                        
-                        let authorName = result.includes.users.first?.name ?? ""
-                        let authorUsername = result.includes.users.first?.username ?? ""
-                        
-                        let tweetToSave = ReturnedTweet(id: result.data.id,
-                                                        text: result.data.text,
-                                                        authorName: authorName,
-                                                        authorUsername: authorUsername)
-                        
-                        completion(.success(tweetToSave))
-                    }
-                    
-                    DispatchQueue.main.async {
-                        // update UI
-                    }
-                    
-                } catch {
-                    print("\(error.localizedDescription)")
-                }
-            }
-        }
-        task.resume()
     }
 }
